requirements.txt
aiohttp>=1.0.5
async-timeout>=1.2.1
certifi>=2017.4.17
chardet>=3.0.3
colorama>=0.3.9
colorlog>=2.10.0
discord.py>=0.16.8,<1.0.0
idna>=2.5
multidict>=2.1.6
olefile>=0.44
Pillow>=4.1.1
pymongo>=3.4.0
pytz>=2017.2
PyYAML>=3.12
requests>=2.17.3
urllib3>=1.21.1
uvloop>=0.8.0
websockets>=3.3
